feat(SectionService): 지하철 구간 관련 로직 분리
