+++ src/main/java/nextstep/subway/ui/LineController.java
@@ -3,6 +3,7 @@ package nextstep.subway.ui;
 import java.net.URI;
 import java.util.List;
 import nextstep.subway.application.LineService;
+import nextstep.subway.application.SectionService;
 import nextstep.subway.dto.LineRequest;
 import nextstep.subway.dto.LineResponse;
 import nextstep.subway.dto.SectionResponse;
@@ -24,9 +25,11 @@ import org.springframework.web.bind.annotation.RestController;
 public class LineController {
 
     private LineService lineService;
+    private SectionService sectionService;
 
-    public LineController(LineService lineService) {
+    public LineController(LineService lineService, SectionService sectionService) {
         this.lineService = lineService;
+        this.sectionService = sectionService;
     }
 
     @PostMapping("/lines")
@@ -58,13 +61,13 @@ public class LineController {
 
     @PostMapping(value = "/lines/{id}/sections", produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<SectionResponse> createSection(@PathVariable Long id, @RequestBody SectionRequest request) {
-        SectionResponse section = lineService.saveSection(id, request);
-        return ResponseEntity.created(URI.create("/" + id + "/sections/" + section.getId())).body(section);
+        SectionResponse section = sectionService.saveSection(id, request);
+        return ResponseEntity.created(URI.create("/sections/" + section.getId())).body(section);
     }
 
     @GetMapping(value = "/lines/{id}/sections", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<List<SectionResponse>> showSections(@PathVariable Long id) {
-        return ResponseEntity.ok().body(lineService.findLineStationsByLineId(id));
+    public ResponseEntity<List<SectionResponse>> showSectionsByLine(@PathVariable Long id) {
+        return ResponseEntity.ok().body(sectionService.findLineStationsByLineId(id));
     }
 
     @ExceptionHandler({DataIntegrityViolationException.class, IllegalArgumentException.class})
