+++ src/main/java/nextstep/subway/ui/StationController.java
@@ -1,6 +1,8 @@
 package nextstep.subway.ui;
 
+import nextstep.subway.application.SectionService;
 import nextstep.subway.application.StationService;
+import nextstep.subway.dto.SectionResponse;
 import nextstep.subway.dto.StationRequest;
 import nextstep.subway.dto.StationResponse;
 import org.springframework.dao.DataIntegrityViolationException;
@@ -15,8 +17,11 @@ import java.util.List;
 public class StationController {
     private StationService stationService;
 
-    public StationController(StationService stationService) {
+    private SectionService sectionService;
+
+    public StationController(StationService stationService, SectionService sectionService) {
         this.stationService = stationService;
+        this.sectionService = sectionService;
     }
 
     @PostMapping("/stations")
@@ -34,6 +39,11 @@ public class StationController {
     public ResponseEntity deleteStation(@PathVariable Long id) {
         stationService.deleteStationById(id);
         return ResponseEntity.noContent().build();
+    }
+
+    @GetMapping(value = "/stations/{id}/sections", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<List<SectionResponse>> showSectionByStation(@PathVariable Long id) {
+        return ResponseEntity.ok().body(sectionService.findLineStationByStationId(id));
     }
 
     @ExceptionHandler(DataIntegrityViolationException.class)
